// front/src/modules/Permission/PermissionValidValue.enum.ts
export enum PermissionValidValue {
  /** Базовые для MMS */
  MMS = 'MMS',

  /** Бронирования */
  MMS_BOOKING = 'MMS_BOOKING',
  MMS_BOOKING_BOOKING = 'MMS_BOOKING_BOOKING',
  MMS_BOOKING_BOOKING_RESTRICTED = 'MMS_BOOKING_BOOKING_PERSON',
  MMS_BOOKING_SPACES = 'MMS_BOOKING_ROOMS',
  MMS_BOOKING_RECORDS = 'MMS_BOOKING_RECORDS',

  /** Конференции */
  MMS_CONFERENCES = 'MMS_CONFERENCES', // todo Добавить в БД
  MMS_CONFERENCES_ACTIVE = 'MMS_CONFERENCES_ACTIVE', // todo Добавить в БД
  MMS_CONFERENCES_HISTORY = 'MMS_CONFERENCES_HISTORY', // todo Добавить в БД

  /** Настройки */
  MMS_SETTINGS = 'MMS_SETTINGS',
  MMS_SETTINGS_USER_PROFILES = 'MMS_SETTINGS_USERPROFILES',
  MMS_SETTINGS_ROOMS_GROUPS = 'MMS_SETTINGS_GROUPSOFROOMS',
  MMS_SETTINGS_RECORD_STORES = 'MMS_SETTINGS_RECORDSTORES',
  MMS_SETTINGS_EXCHANGE = 'MMS_SETTINGS_EXCHANGE',

  MMS_USERS = 'MMS_USERS', // не используется
  MMS_USERS_USERS = 'MMS_USERS_USERS', // не используется
  MMS_USERS_ROLES = 'MMS_USERS_ROLES', // не используется
  MMS_USERS_LOGS = 'MMS_USERS_LOGS', // не используется

  /** Сервера CMS */
  MMS_CMS = 'MMS_CMS',
  MMS_CMS_SERVERS = 'MMS_CMS_SERVERS',
  MMS_CMS_SERVERS_GROUPS = 'MMS_CMS_GROUPS',

  /** Отчеты */
  MMS_REPORTS = 'MMS_REPORTS',
  MMS_REPORTS_REPORTS = 'MMS_REPORTS_REPORTS',
  MMS_REPORTS_REPORTS_DISTRIBUTION = 'MMS_REPORTS_DISTRIBUTIONOFREPORTS',
}